dev 20201103
fix home tab icon on widget update
merge home page grid item class
merge bug fix from vietwin
change service navigate to service route

+++ lib/features/routes/home/data/entity/game_entity.dart
@@ -43,7 +43,7 @@ extension GameEntityExtension on GameEntity {
   }
 
   String get imageUrl => '/images/index/gamelist/$id.jpg';
-  bool isLongText(int limit) => (Global.lang == 'zh')
+  bool isLongText(double limit) => (Global.lang == 'zh')
       ? cname.countLength > limit
       : ename.countLength > limit;
 }
